# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface PortfolioHolding {
  id: string;
  card: string;
  grade: string;
  cost: number;
  value: number;
  gain: number;
  gainPct: number;
  purchaseDate?: string | null;
  createdAt: string;
}

export interface CreatePortfolioHolding {
  card: string;
  grade: string;
  cost: number;
  value: number;
  purchaseDate?: string;
}

export interface UpdatePortfolioHolding {
  grade?: string;
  value?: number;
  cost?: number;
}

export interface ScanResult {
  id: string;
  cardName: string;
  year?: string | null;
  setName?: string | null;
  parallel?: string | null;
  askingPrice?: number | null;
  shipping?: number | null;
  estValue?: number | null;
  estGrade?: string | null;
  gradeRange?: string | null;
  probability?: number | null;
  roi?: number | null;
  recommendedAction?: string | null;
  imageQualityScore?: number | null;
  createdAt: string;
}

export interface CreateScanResult {
  cardName: string;
  year?: string;
  setName?: string;
  parallel?: string;
  askingPrice?: number;
  shipping?: number;
  estValue?: number;
  estGrade?: string;
  gradeRange?: string;
  probability?: number;
  roi?: number;
  recommendedAction?: string;
  imageQualityScore?: number;
}

export interface AnalyzeListingRequest {
  listingUrl: string;
  askingPrice?: number;
  shipping?: number;
}

export interface ConditionDetail {
  score: number;
  status: string;
}

export interface MarketComps {
  raw: number[];
  psa8: number[];
  psa9: number[];
  psa10: number[];
}

export type AnalysisResultCondition = { [key: string]: ConditionDetail };

export interface AnalysisResult {
  cardName: string;
  player: string;
  year: string;
  setName: string;
  cardNumber: string;
  parallel: string;
  estGrade: string;
  gradeRange: string;
  probability: number;
  estValue: number;
  roi: number;
  recommendedAction: string;
  imageQualityScore: number;
  condition: AnalysisResultCondition;
  notes: string[];
  marketComps: MarketComps;
}

export type GradingSubmissionGrader =
  (typeof GradingSubmissionGrader)[keyof typeof GradingSubmissionGrader];

export const GradingSubmissionGrader = {
  PSA: "PSA",
  BGS: "BGS",
  SGC: "SGC",
  CGC: "CGC",
} as const;

export type GradingSubmissionStatus =
  (typeof GradingSubmissionStatus)[keyof typeof GradingSubmissionStatus];

export const GradingSubmissionStatus = {
  pending: "pending",
  "in-grading": "in-grading",
  graded: "graded",
  shipped: "shipped",
  completed: "completed",
} as const;

export interface GradingSubmission {
  id: string;
  cardName: string;
  grader: GradingSubmissionGrader;
  serviceLevel: string;
  declaredValue?: number | null;
  submittedDate?: string | null;
  returnedDate?: string | null;
  certNumber?: string | null;
  status: GradingSubmissionStatus;
  gradeReceived?: string | null;
  notes?: string | null;
  createdAt: string;
}

export type CreateGradingSubmissionGrader =
  (typeof CreateGradingSubmissionGrader)[keyof typeof CreateGradingSubmissionGrader];

export const CreateGradingSubmissionGrader = {
  PSA: "PSA",
  BGS: "BGS",
  SGC: "SGC",
  CGC: "CGC",
} as const;

export interface CreateGradingSubmission {
  cardName: string;
  grader: CreateGradingSubmissionGrader;
  serviceLevel: string;
  declaredValue?: number;
  submittedDate?: string;
  notes?: string;
}

export type UpdateGradingSubmissionStatus =
  (typeof UpdateGradingSubmissionStatus)[keyof typeof UpdateGradingSubmissionStatus];

export const UpdateGradingSubmissionStatus = {
  pending: "pending",
  "in-grading": "in-grading",
  graded: "graded",
  shipped: "shipped",
  completed: "completed",
} as const;

export interface UpdateGradingSubmission {
  status?: UpdateGradingSubmissionStatus;
  gradeReceived?: string;
  certNumber?: string;
  returnedDate?: string;
  declaredValue?: number;
  notes?: string;
  serviceLevel?: string;
}

export type WantlistItemPriority =
  (typeof WantlistItemPriority)[keyof typeof WantlistItemPriority];

export const WantlistItemPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface WantlistItem {
  id: string;
  cardName: string;
  targetGrade: string;
  maxPrice: number;
  priority: WantlistItemPriority;
  notes?: string | null;
  acquired: boolean;
  createdAt: string;
}

export type CreateWantlistItemPriority =
  (typeof CreateWantlistItemPriority)[keyof typeof CreateWantlistItemPriority];

export const CreateWantlistItemPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateWantlistItem {
  cardName: string;
  targetGrade: string;
  maxPrice: number;
  priority?: CreateWantlistItemPriority;
  notes?: string;
}

export type UpdateWantlistItemPriority =
  (typeof UpdateWantlistItemPriority)[keyof typeof UpdateWantlistItemPriority];

export const UpdateWantlistItemPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateWantlistItem {
  cardName?: string;
  targetGrade?: string;
  maxPrice?: number;
  priority?: UpdateWantlistItemPriority;
  notes?: string;
  acquired?: boolean;
}

export interface PortfolioSnapshot {
  id: string;
  /** ISO date string e.g. 2026-05-01 */
  snapshotDate: string;
  /** Total portfolio value in USD at time of snapshot */
  totalValue: number;
  createdAt: string;
}