# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: portfolio
    description: Portfolio holdings management
  - name: scans
    description: Scan history and AI analysis
  - name: grading
    description: Grading submission tracking
  - name: wantlist
    description: Cards you want to acquire
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /portfolio:
    get:
      operationId: listPortfolioHoldings
      tags: [portfolio]
      summary: List portfolio holdings
      description: Returns all portfolio holdings for the authenticated user
      responses:
        "200":
          description: List of holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioHolding"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPortfolioHolding
      tags: [portfolio]
      summary: Add a portfolio holding
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortfolioHolding"
      responses:
        "201":
          description: Created holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHolding"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/{id}:
    put:
      operationId: updatePortfolioHolding
      tags: [portfolio]
      summary: Update a portfolio holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortfolioHolding"
      responses:
        "200":
          description: Updated holding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioHolding"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePortfolioHolding
      tags: [portfolio]
      summary: Remove a portfolio holding
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /portfolio/history:
    get:
      operationId: getPortfolioHistory
      tags: [portfolio]
      summary: Get portfolio value history
      description: Returns daily portfolio value snapshots for the authenticated user. Creates a snapshot for today on demand if missing.
      responses:
        "200":
          description: List of daily value snapshots ordered by date ascending
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioSnapshot"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /grading-submissions:
    get:
      operationId: listGradingSubmissions
      tags: [grading]
      summary: List grading submissions
      description: Returns all grading submissions for the authenticated user, newest first
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GradingSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createGradingSubmission
      tags: [grading]
      summary: Log a grading submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGradingSubmission"
      responses:
        "201":
          description: Created submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /grading-submissions/{id}:
    put:
      operationId: updateGradingSubmission
      tags: [grading]
      summary: Update a grading submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGradingSubmission"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradingSubmission"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGradingSubmission
      tags: [grading]
      summary: Remove a grading submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wantlist:
    get:
      operationId: listWantlistItems
      tags: [wantlist]
      summary: List wantlist items
      description: Returns all wantlist items for the authenticated user
      responses:
        "200":
          description: List of wantlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WantlistItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWantlistItem
      tags: [wantlist]
      summary: Add a card to wantlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWantlistItem"
      responses:
        "201":
          description: Created wantlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WantlistItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wantlist/{id}:
    put:
      operationId: updateWantlistItem
      tags: [wantlist]
      summary: Update a wantlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWantlistItem"
      responses:
        "200":
          description: Updated wantlist item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WantlistItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWantlistItem
      tags: [wantlist]
      summary: Remove a wantlist item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /scans:
    get:
      operationId: listScanResults
      tags: [scans]
      summary: List scan history
      description: Returns scan history for the authenticated user, newest first
      responses:
        "200":
          description: List of scan results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createScanResult
      tags: [scans]
      summary: Save a scan result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanResult"
      responses:
        "201":
          description: Saved scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyze-listing:
    post:
      operationId: analyzeListing
      tags: [scans]
      summary: Analyze a card listing with AI
      description: Uses AI to assess grading potential, condition, and ROI from a marketplace listing URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeListingRequest"
      responses:
        "200":
          description: AI analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    PortfolioHolding:
      type: object
      properties:
        id:
          type: string
        card:
          type: string
        grade:
          type: string
        cost:
          type: integer
        value:
          type: integer
        gain:
          type: integer
        gainPct:
          type: number
        purchaseDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - card
        - grade
        - cost
        - value
        - gain
        - gainPct
        - createdAt

    CreatePortfolioHolding:
      type: object
      properties:
        card:
          type: string
        grade:
          type: string
        cost:
          type: integer
        value:
          type: integer
        purchaseDate:
          type: string
      required:
        - card
        - grade
        - cost
        - value

    UpdatePortfolioHolding:
      type: object
      properties:
        grade:
          type: string
        value:
          type: integer
        cost:
          type: integer
      minProperties: 1

    ScanResult:
      type: object
      properties:
        id:
          type: string
        cardName:
          type: string
        year:
          type: string
          nullable: true
        setName:
          type: string
          nullable: true
        parallel:
          type: string
          nullable: true
        askingPrice:
          type: number
          nullable: true
        shipping:
          type: number
          nullable: true
        estValue:
          type: integer
          nullable: true
        estGrade:
          type: string
          nullable: true
        gradeRange:
          type: string
          nullable: true
        probability:
          type: integer
          nullable: true
        roi:
          type: number
          nullable: true
        recommendedAction:
          type: string
          nullable: true
        imageQualityScore:
          type: integer
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - cardName
        - createdAt

    CreateScanResult:
      type: object
      properties:
        cardName:
          type: string
        year:
          type: string
        setName:
          type: string
        parallel:
          type: string
        askingPrice:
          type: number
        shipping:
          type: number
        estValue:
          type: integer
        estGrade:
          type: string
        gradeRange:
          type: string
        probability:
          type: integer
        roi:
          type: number
        recommendedAction:
          type: string
        imageQualityScore:
          type: integer
      required:
        - cardName

    AnalyzeListingRequest:
      type: object
      properties:
        listingUrl:
          type: string
        askingPrice:
          type: number
        shipping:
          type: number
      required:
        - listingUrl

    ConditionDetail:
      type: object
      properties:
        score:
          type: number
        status:
          type: string
      required:
        - score
        - status

    MarketComps:
      type: object
      properties:
        raw:
          type: array
          items:
            type: number
        psa8:
          type: array
          items:
            type: number
        psa9:
          type: array
          items:
            type: number
        psa10:
          type: array
          items:
            type: number
      required:
        - raw
        - psa8
        - psa9
        - psa10

    AnalysisResult:
      type: object
      properties:
        cardName:
          type: string
        player:
          type: string
        year:
          type: string
        setName:
          type: string
        cardNumber:
          type: string
        parallel:
          type: string
        estGrade:
          type: string
        gradeRange:
          type: string
        probability:
          type: integer
        estValue:
          type: integer
        roi:
          type: number
        recommendedAction:
          type: string
        imageQualityScore:
          type: integer
        condition:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/ConditionDetail"
        notes:
          type: array
          items:
            type: string
        marketComps:
          $ref: "#/components/schemas/MarketComps"
      required:
        - cardName
        - player
        - year
        - setName
        - cardNumber
        - parallel
        - estGrade
        - gradeRange
        - probability
        - estValue
        - roi
        - recommendedAction
        - imageQualityScore
        - condition
        - notes
        - marketComps

    GradingSubmission:
      type: object
      properties:
        id:
          type: string
        cardName:
          type: string
        grader:
          type: string
          enum: [PSA, BGS, SGC, CGC]
        serviceLevel:
          type: string
        declaredValue:
          type: integer
          nullable: true
        submittedDate:
          type: string
          nullable: true
        returnedDate:
          type: string
          nullable: true
        certNumber:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in-grading, graded, shipped, completed]
        gradeReceived:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - cardName
        - grader
        - serviceLevel
        - status
        - createdAt

    CreateGradingSubmission:
      type: object
      properties:
        cardName:
          type: string
        grader:
          type: string
          enum: [PSA, BGS, SGC, CGC]
        serviceLevel:
          type: string
        declaredValue:
          type: integer
        submittedDate:
          type: string
        notes:
          type: string
      required:
        - cardName
        - grader
        - serviceLevel

    UpdateGradingSubmission:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in-grading, graded, shipped, completed]
        gradeReceived:
          type: string
        certNumber:
          type: string
        returnedDate:
          type: string
        declaredValue:
          type: integer
        notes:
          type: string
        serviceLevel:
          type: string
      minProperties: 1

    WantlistItem:
      type: object
      properties:
        id:
          type: string
        cardName:
          type: string
        targetGrade:
          type: string
        maxPrice:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
          nullable: true
        acquired:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - cardName
        - targetGrade
        - maxPrice
        - priority
        - acquired
        - createdAt

    CreateWantlistItem:
      type: object
      properties:
        cardName:
          type: string
        targetGrade:
          type: string
        maxPrice:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
      required:
        - cardName
        - targetGrade
        - maxPrice

    UpdateWantlistItem:
      type: object
      properties:
        cardName:
          type: string
        targetGrade:
          type: string
        maxPrice:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        notes:
          type: string
        acquired:
          type: boolean
      minProperties: 1

    PortfolioSnapshot:
      type: object
      properties:
        id:
          type: string
        snapshotDate:
          type: string
          description: ISO date string e.g. 2026-05-01
        totalValue:
          type: integer
          description: Total portfolio value in USD at time of snapshot
        createdAt:
          type: string
      required:
        - id
        - snapshotDate
        - totalValue
        - createdAt