# lib/api-zod/src/generated/types/createGradingSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateGradingSubmissionGrader } from "./createGradingSubmissionGrader";

export interface CreateGradingSubmission {
  cardName: string;
  grader: CreateGradingSubmissionGrader;
  serviceLevel: string;
  declaredValue?: number;
  submittedDate?: string;
  notes?: string;
}