# lib/api-zod/src/generated/types/createWantlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateWantlistItemPriority } from "./createWantlistItemPriority";

export interface CreateWantlistItem {
  cardName: string;
  targetGrade: string;
  maxPrice: number;
  priority?: CreateWantlistItemPriority;
  notes?: string;
}