# lib/api-zod/src/generated/types/gradingSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GradingSubmissionGrader } from "./gradingSubmissionGrader";
import type { GradingSubmissionStatus } from "./gradingSubmissionStatus";

export interface GradingSubmission {
  id: string;
  cardName: string;
  grader: GradingSubmissionGrader;
  serviceLevel: string;
  declaredValue?: number | null;
  submittedDate?: string | null;
  returnedDate?: string | null;
  certNumber?: string | null;
  status: GradingSubmissionStatus;
  gradeReceived?: string | null;
  notes?: string | null;
  createdAt: string;
}