# lib/api-zod/src/generated/types/updateGradingSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateGradingSubmissionStatus } from "./updateGradingSubmissionStatus";

export interface UpdateGradingSubmission {
  status?: UpdateGradingSubmissionStatus;
  gradeReceived?: string;
  certNumber?: string;
  returnedDate?: string;
  declaredValue?: number;
  notes?: string;
  serviceLevel?: string;
}